+++ util1.c
@@ -227,329 +227,6 @@ int full_write(int desc, const char *ptr, size_t len)
 	return total_written;
 }
 
-/**
- * Read @p len bytes at @p ptr from descriptor @p desc, retrying if
- * interrupted.
- *
- * @retval >0 the actual number of bytes read
- *
- * @retval 0 for EOF
- *
- * @retval <0 for an error.
- *
- * Derived from GNU C's cccp.c. */
-static int safe_read(int desc, char *ptr, size_t len)
-{
-	int n_chars;
-
-	if (len == 0)
-		return len;
-
-	do {
-		n_chars = read(desc, ptr, len);
-	} while (n_chars < 0 && errno == EINTR);
-
-	return n_chars;
-}
-
-/* Remove existing file @dest and reopen, creating a new file with @mode */
-static int unlink_and_reopen(const char *dest, mode_t mode)
-{
-	int ofd;
-
-	if (robust_unlink(dest) && errno != ENOENT) {
-		int save_errno = errno;
-		rsyserr(FERROR_XFER, errno, "unlink %s", full_fname(dest));
-		errno = save_errno;
-		return -1;
-	}
-
-#ifdef SUPPORT_XATTRS
-	if (preserve_xattrs)
-		mode |= S_IWUSR;
-#endif
-	mode &= INITACCESSPERMS;
-	/* Use vfs_open_at so the create/truncate goes through a secure
-	 * parent dirfd in the daemon-no-chroot deployment. Otherwise
-	 * an attacker could swap a parent component with a symlink in
-	 * the window between robust_unlink (which uses vfs_unlink_at,
-	 * already secure) and the create here, and redirect the new
-	 * file outside the module. */
-	if ((ofd = vfs_open_at(dest, O_WRONLY | O_CREAT | O_TRUNC | O_EXCL, mode)) < 0) {
-		int save_errno = errno;
-		rsyserr(FERROR_XFER, save_errno, "open %s", full_fname(dest));
-		errno = save_errno;
-		return -1;
-	}
-	return ofd;
-}
-
-/* Copy contents of file @source to file @dest with mode @mode.
- *
- * If @tmpfilefd is < 0, copy_file unlinks @dest and then opens a new
- * file with name @dest.
- *
- * Otherwise, copy_file writes to and closes the provided file
- * descriptor.
- *
- * In either case, if --xattrs are being preserved, the dest file will
- * have its xattrs set from the source file.
- *
- * This is used in conjunction with the --temp-dir, --backup, and
- * --copy-dest options. */
-int copy_file(const char *source, const char *dest, int tmpfilefd, mode_t mode)
-{
-	int ifd, ofd;
-	char buf[1024 * 8];
-	int len;   /* Number of bytes read into `buf'. */
-	OFF_T prealloc_len = 0, offset = 0;
-
-	/* For any hardened (non-chrooted) receiver, route the source open through
-	 * vfs_resolve_open so a parent-symlink on the source path (e.g.
-	 * --copy-dest=cd where cd is a symlink to an outside directory) cannot
-	 * redirect the read to a file the attacker should not see.  Plain
-	 * vfs_open_nofollow only refuses a final-component symlink; parents are
-	 * still followed.  (An absolute source is operator-trusted -- e.g. an
-	 * absolutized basis dir -- and uses vfs_open_nofollow.) */
-	if (vfs_relpath_active() && source && *source && source[0] != '/')
-		ifd = vfs_resolve_open(NULL, source, O_RDONLY | O_NOFOLLOW, 0);
-	else
-		ifd = vfs_open_nofollow(source, O_RDONLY);
-	if (ifd < 0) {
-		int save_errno = errno;
-		rsyserr(FERROR_XFER, errno, "open %s", full_fname(source));
-		errno = save_errno;
-		return -1;
-	}
-
-	if (tmpfilefd >= 0) {
-		ofd = tmpfilefd;
-	} else {
-		ofd = unlink_and_reopen(dest, mode);
-		if (ofd < 0) {
-			int save_errno = errno;
-			close(ifd);
-			errno = save_errno;
-			return -1;
-		}
-	}
-
-#ifdef SUPPORT_PREALLOCATION
-	if (preallocate_files) {
-		STRUCT_STAT srcst;
-
-		/* Try to preallocate enough space for file's eventual length.  Can
-		 * reduce fragmentation on filesystems like ext4, xfs, and NTFS. */
-		if (vfs_fstat(ifd, &srcst) < 0)
-			rsyserr(FWARNING, errno, "fstat %s", full_fname(source));
-		else if (srcst.st_size > 0) {
-			prealloc_len = vfs_fallocate(ofd, 0, srcst.st_size);
-			if (prealloc_len < 0)
-				rsyserr(FWARNING, errno, "vfs_fallocate %s", full_fname(dest));
-		}
-	}
-#endif
-
-	while ((len = safe_read(ifd, buf, sizeof buf)) > 0) {
-		if (full_write(ofd, buf, len) < 0) {
-			int save_errno = errno;
-			rsyserr(FERROR_XFER, errno, "write %s", full_fname(dest));
-			close(ifd);
-			close(ofd);
-			errno = save_errno;
-			return -1;
-		}
-		offset += len;
-	}
-
-	if (len < 0) {
-		int save_errno = errno;
-		rsyserr(FERROR_XFER, errno, "read %s", full_fname(source));
-		close(ifd);
-		close(ofd);
-		errno = save_errno;
-		return -1;
-	}
-
-	/* Source file might have shrunk since we fstatted it.
-	 * Cut off any extra preallocated zeros from dest file. */
-	if (offset < prealloc_len) {
-#ifdef HAVE_FTRUNCATE
-		/* If we fail to truncate, the dest file may be wrong, so we
-		 * must trigger the "partial transfer" error. */
-		if (vfs_ftruncate(ofd, offset) < 0)
-			rsyserr(FERROR_XFER, errno, "ftruncate %s", full_fname(dest));
-#else
-		rprintf(FERROR_XFER, "no ftruncate for over-long pre-alloc: %s", full_fname(dest));
-#endif
-	}
-
-	if (do_fsync && fsync(ofd) < 0) {
-		int save_errno = errno;
-		rsyserr(FERROR, errno, "fsync failed on %s", full_fname(dest));
-		close(ofd);
-		close(ifd);	/* ifd is held open until after the xattr copy below */
-		errno = save_errno;
-		return -1;
-	}
-
-#ifdef SUPPORT_XATTRS
-	/* Read the source xattrs through the held source fd (ifd) and set them
-	 * through ofd while both are still held, so a parent-symlink race can't
-	 * redirect the read out of tree or the write onto a file outside it. */
-	if (preserve_xattrs)
-		copy_xattrs(source, ifd, dest, ofd);
-#endif
-
-	if (close(ifd) < 0) {
-		rsyserr(FWARNING, errno, "close failed on %s",
-			full_fname(source));
-	}
-
-	if (close(ofd) < 0) {
-		int save_errno = errno;
-		rsyserr(FERROR_XFER, errno, "close failed on %s", full_fname(dest));
-		errno = save_errno;
-		return -1;
-	}
-
-	return 0;
-}
-
-/* MAX_RENAMES should be 10**MAX_RENAMES_DIGITS */
-#define MAX_RENAMES_DIGITS 3
-#define MAX_RENAMES 1000
-
-/**
- * Robust unlink: some OS'es (HPUX) refuse to unlink busy files, so
- * rename to <path>/.rsyncNNN instead.
- *
- * Note that successive rsync runs will shuffle the filenames around a
- * bit as long as the file is still busy; this is because this function
- * does not know if the unlink call is due to a new file coming in, or
- * --delete trying to remove old .rsyncNNN files, hence it renames it
- * each time.
- **/
-int robust_unlink(const char *fname)
-{
-#ifndef ETXTBSY
-	return vfs_unlink_at(fname);
-#else
-	static int counter = 1;
-	int rc, pos, start;
-	char path[MAXPATHLEN];
-
-	rc = vfs_unlink_at(fname);
-	if (rc == 0 || errno != ETXTBSY)
-		return rc;
-
-	if ((pos = strlcpy(path, fname, MAXPATHLEN)) >= MAXPATHLEN)
-		pos = MAXPATHLEN - 1;
-
-	while (pos > 0 && path[pos-1] != '/')
-		pos--;
-	pos += strlcpy(path+pos, ".rsync", MAXPATHLEN-pos);
-
-	if (pos > (MAXPATHLEN-MAX_RENAMES_DIGITS-1)) {
-		errno = ETXTBSY;
-		return -1;
-	}
-
-	/* start where the last one left off to reduce chance of clashes */
-	start = counter;
-	do {
-		snprintf(&path[pos], MAX_RENAMES_DIGITS+1, "%03d", counter);
-		if (++counter >= MAX_RENAMES)
-			counter = 1;
-	} while (access(path, 0) == 0 && counter != start);
-
-	if (INFO_GTE(MISC, 1)) {
-		rprintf(FWARNING, "renaming %s to %s because of text busy\n",
-			fname, path);
-	}
-
-	/* maybe we should return rename()'s exit status? Nah. */
-	if (vfs_rename_at(fname, path) != 0) {
-		errno = ETXTBSY;
-		return -1;
-	}
-	return 0;
-#endif
-}
-
-/* Returns 0 on successful rename, 1 if we successfully copied the file
- * across filesystems, -2 if copy_file() failed, and -1 on other errors.
- * If partialptr is not NULL and we need to do a copy, copy the file into
- * the active partial-dir instead of over the destination file. */
-int robust_rename(const char *from, const char *to, const char *partialptr,
-		  int mode, struct file_struct *file)
-{
-	int tries = 4;
-
-	/* A resumed in-place partial-dir transfer might call us with from and
-	 * to pointing to the same buf if the transfer failed yet again. */
-	if (from == to)
-		return 0;
-
-	while (tries--) {
-		/* tmp -> final usually live in the entry's own dir: rename via the
-		 * held dir fd when both do, else the full-path wrapper. */
-		int ofd = vfs_cached_dirfd(from, file);
-		int nfd = vfs_cached_dirfd(to, file);
-		int rr;
-		if (ofd >= 0 && nfd >= 0) {
-			const char *os = strrchr(from, '/');
-			const char *ns = strrchr(to, '/');
-			rr = vfs_rename_atfd(ofd, os ? os + 1 : from, nfd, ns ? ns + 1 : to);
-		} else
-			rr = vfs_rename_at(from, to);
-		if (rr == 0)
-			return 0;
-
-		switch (errno) {
-#ifdef ETXTBSY
-		case ETXTBSY:
-			if (robust_unlink(to) != 0) {
-				errno = ETXTBSY;
-				return -1;
-			}
-			errno = ETXTBSY;
-			break;
-#endif
-		case EXDEV: {
-			int save = operator_path_resolve, rc;
-			if (partialptr) {
-				if (!handle_partial_dir(partialptr,PDIR_CREATE))
-					return -2;
-				to = partialptr;
-			}
-			/* Cross-fs fallback: copy then unlink.  An absolute --temp-dir
-			 * source / --partial-dir dest is an operator path whose parents
-			 * do_open_at()/do_unlink_at() would otherwise follow via plain libc
-			 * -- confine them through the ownership walk so a raced parent
-			 * symlink can't redirect the dest-write or the source-unlink out of
-			 * the module.  copy_file already confines the source READ; a
-			 * relative in-module path stays on the secure_relative_open arm, so
-			 * only flip the flag for an absolute (operator) path. */
-			if (*to == '/')
-				operator_path_resolve = 1;
-			rc = copy_file(from, to, -1, mode);
-			operator_path_resolve = save;
-			if (rc != 0)
-				return -2;
-			if (*from == '/')
-				operator_path_resolve = 1;
-			vfs_unlink_at(from);
-			operator_path_resolve = save;
-			return 1;
-		}
-		default:
-			return -1;
-		}
-	}
-	return -1;
-}
-
 static pid_t all_pids[10];
 static int num_pids;
